# backend/docs/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Revive AI API
  description: |
    AI-Powered Reputation Intelligence & Customer Recovery Platform
    
    ## Overview
    Revive AI helps businesses monitor public reviews, analyze customer conversations,
    predict churn risks, and take automated recovery actions to improve ratings,
    retention, and revenue.
    
    ## Authentication
    All endpoints require JWT Bearer token authentication (except public endpoints).
    Include the token in the Authorization header: `Bearer <token>`
  version: 1.0.0
  contact:
    name: Revive AI Support
    email: support@revive-ai.com
    url: https://revive-ai.com/support
  license:
    name: Proprietary License
    url: https://revive-ai.com/license
  termsOfService: https://revive-ai.com/terms
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://staging-api.revive-ai.com
    description: Staging server
  - url: https://api.revive-ai.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              example: validation_error
            message:
              type: string
              example: Request validation failed
            status_code:
              type: integer
              example: 422
            request_id:
              type: string
              format: uuid
            timestamp:
              type: number
              format: float
            path:
              type: string
              example: /api/v1/reviews
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - authentication
      summary: Register new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - organization_name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                organization_name:
                  type: string
      responses:
        '201':
          description: Account created successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/login:
    post:
      tags:
        - authentication
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
                  expires_in:
                    type: integer
                    example: 3600
...